+++ previous-versions/files/RELEASES.md
@@ -575,3 +575,13 @@ Based on:
 - Speakeasy CLI 1.53.0 (2.58.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [PyPI v0.33.1] https://pypi.org/project/codat-files/0.33.1 - files
+
+## 2023-09-01 15:58:32
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.77.1 (2.91.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.34.0] previous-versions/files
+### Releases
+- [PyPI v0.34.0] https://pypi.org/project/codat-files/0.34.0 - previous-versions/files